container/requirements.txt
torch
transformers
tokenizers
pandas
BioPython